+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2017-7-26
  * Add support for SVG tags -- this BREAKS any substitutions using the same name as an SVG tag
- * Add larceny ("l") namespace for ambiguous tags
+ * Add larceny ("l") prefix for ambiguous tags
  * Add start of a conversion script to help find SVG-shadowed substitutions. It doesn't actually work as-is -- you need to make some internal Larceny modules accessible and change the template directory inside the script. Not sure how/if to address this for now, because there are only a couple of people who are ever going to use this script.
 
 ## 2017-7-21